pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "shepherd_wsrv"
description = "Web-Server & -API for the Shepherd-Testbed"
keywords = ["testbed", "beaglebone", "pru", "batteryless", "energyharvesting", "solar"]

authors = [{name = "Ingmar Splitt", email = "ingmar.splitt@tu-dresden.de"},]
maintainers = [{name = "Ingmar Splitt", email = "ingmar.splitt@tu-dresden.de"},]

readme = {file = "README.md", content-type = "text/markdown"}
license = {file = "LICENSE"}
dynamic = ["version"]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Natural Language :: English"
]

requires-python = ">=3.10"
dependencies = [
    "pandas",  # todo: maybe ditch, just use numpy, if needed at all
    "pyaml",
    "fastapi",
    "fastapi-mail",
    "typer[all]",
    "uvicorn[standard]",
    "python-multipart",
    "pydantic[email]",
    "beanie",
    "fastui",
    "python-secrets",
    "shepherd_core==2023.12.1",
    "python-decouple",
#   "shepherd_herd", # TODO
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "pyright",
    "pandas-stubs",  # for pyright with pandas
    # as long as pandas has to type hints
    "bcrypt", # for salt-creation
]
test = [
    "pytest",
    "pytest-click",
    "coverage",
]

[project.urls]
Documentation = "https://github.com/orgua/shepherd_webservice/blob/main/README.md"
Issues = "https://github.com/orgua/shepherd_webservice/issues"
Source = "https://github.com/orgua/shepherd_webservice"

[project.scripts]
shepherd-wsrv = "shepherd_wsrv.cli:cli"

[tool.setuptools]
platforms = ["unix", "linux", "osx", "cygwin", "win32", "win64"]
zip-safe = true

[tool.setuptools.package-dir]
shepherd_wsrv = "shepherd_wsrv"

[tool.setuptools.dynamic]
version = {attr = "shepherd_wsrv.__version__"}
#version = {file = "./shepherd_wsrv/config.py:__version__"}  # TODO: does not work

[tool.pyright]
root = "./"
include = ['./sratch_models',]  # TODO
pythonVersion = "3.10"
pythonPlatform = "All"
reportMissingParameterType = true

[tool.pytest.ini_options]
addopts = ["-vvv"]

[tool.coverage.run]
source = ["shepherd_wsrv"]
# TODO: add something like "--cov --cov-report html --cov-report term-missing --cov-fail-under 95"
